+++ fast_barcode_scanner/example/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'scanning_screen/scanner_screen.dart';
+import 'scanning_screen/scanning_screen.dart';
 
 void main() {
   runApp(const MaterialApp(home: ScanningScreen()));